fix(example): Move utils.h to dedicated directory
