+++ example/CMakeLists.txt
@@ -2,6 +2,7 @@ add_executable(example
         main.cpp
         src/user/user_controller.cpp
         src/user/user_module.cpp
+        src/utils/utils.cpp
         src/main_controller.cpp
         src/main_module.cpp
 )
