+++ example/include/main_module.h
@@ -1,7 +1,7 @@
 #ifndef EXAMPLE_MAIN_MODULE_H
 #define EXAMPLE_MAIN_MODULE_H
 
-#include "utils.h"
+#include "utils/utils.h"
 
 namespace example {
     class main_module: public module {
