+++ example/main.cpp
@@ -1,6 +1,6 @@
 #include <functional>
 
-#include "utils.h"
+#include "utils/utils.h"
 
 #include "main_module.h"
