+++ example/src/user/user_module.cpp
@@ -1,4 +1,5 @@
 #include "user/user_module.h"
+#include "user/user_controller.h"
 
 namespace example {
     user_module::user_module() {
